chore(profiler): span link independent from context provider
We make sure that the span link mechanism in the profiler is independent
from the context provider, as this can change with calls to `configure`.
Instead we rely on the core event hub to emit an event when a span is
activated so that the profiler can perform the linking with the
corresponding thread.