chore(profiling): remove leak unit test for now
It's not currently working as intended and not clear that there's a path
forward for that test that won't be either too sensitve (flaky or tied
to implementation details) or not sensitive enough (don't actually catch
the problem). Remove it for now and rely on the relenv to evaluate the
fix.