fix(profiling): reduce lock profiler log spam (#11377)
The lock profiler would log a warning if it couldn't determine a
name for a lock, and it would try determining a name multiple times for
the same lock. This lead to excessive log spam. Downgrade this to a
debug log and only try to determine the name once.

(cherry picked from commit 079e4789ccd86b1aa567238d525ba45c1cedce56)