chore(tracing): fix inferred base service for pytest command ran with --ddtrace optional argument (#11394)
# Motivation

Add special case for `--ddtrace` pytest argument, that was causing CI tests to have a different service name than local testing. CI testing adds the `--ddtrace` arg to the command. Previously the code would skip any args that started with `-` as well as the following arg, but that should not be the case for this arg.