refactor(profiler): tweak sample string arena default size
Given that each frame gets a file and function name, and we default to
up to 64 frames, it seems reasonable to have more than 1024 bytes for
strings. I set 1024 when I thought we just had a few tags. We can
probably end up with a more than 64KiB for deep stacks with long
filenames, but I for now I'll lean towards something smaller and be
optimistic that we don't tend to need more. Would be good to measure
what we actually tend to use.