refactor(profiling): tweak chunk size again and document StringArena
I downloaded a bunch of Python reliability enviornment profiles and
computed, for each sample, the total size in bytes of all the strings
associated with the sample from label value strings and funciton and
file names. The largest I saw was ~10000 bytes. So I think 16KiB is a
reasonble default chunk size for the arena. Updated accordingly.

I also documented the StringArena. Specifically, what it's even for and
what the members do. I also aliased the "chunk" type to make it easier
to read the struct definition.