chore(tracer): add ProductSpanProcessors
It has become apparent that there are two different types of
SpanProcessor that are utilized in the tracer: tracing related span
processors and additional product processors. As there is no current
separation of these types, the product logic built on top of tracing is
strongly coupled.

I believe it makes sense to decouple these concepts explicitly in the
tracer. There is too much ASM code in the tracer module because there
does not exist an interface for ASM to interface with the tracer.

LLMObs suffers a similar problem: a traceprocessor is attached via the
public `configure()` API but this suffers from conflicts with end-user
usage of the same API.