fix(profiling): get rid of leaky string table (#11612)
Profiler samples typically contain several strings, for sample labels
(task name, endpoint/resource, etc) and frame info. Those strings tend
to be repeated. We build samples across several API calls, from Python,
Cython, C Python extensions, and C++ code, making it hard to reason
about the lifetimes of the strings we get. So, it's easier to just copy
the strings we need while building a sample. And it's preferably to only
store one copy. So, our libdatadog wrapper uses a string table, which
addresses both of these concerns. However, nothing is ever removed from
the table. And some things like task names can have new values each
time, meaning we have unbounded memory growth.

I previously tried to fix this by periodically clearing the string
table. However, this comes with its own new problems. We can construct
samples in parallel with profile uploading. So if we try to clear the
table when we upload the profile, or really any other time, we need to
make sure in-progress samples keep their strings around until the sample
is flushed to libdatadog. And we need to handle forking so we don't leak
a whole string table on fork.

It turns out we don't really need that string table at all. Libdatadog
already interns strings once the sample makes it into the libdatadog
profile. That solves the duplicate string problem. We still want to copy
strings while building the sample, but we can just do that on a
per-sample basis, and the strings only need to be alive while we build a
sample. We can use a small amount of scratch space for each sample to
copy strings. We will likely only have a few samples in flight at any
time, and we already pool the samples. So we can reuse this scratch
memory effectively. We'll only use as memory in proportion to the
number of samples being built concurrently.

So, this commit implements a simple string arena for each sample object
to use. We may want to tweak the default buffer size and how much memory
we keep around when resetting the arena. We could even deduplicate,
since we might see some duplicates per-sample from frame info.

PROF-10824