chore(di): accurate function line mapping (#11607)
We make the line-to-function mapping more accurate by capturing the code
objects exported by a module before the module is executed. This gives
us the full code object content of the model, which matches the source
more closely than the executed module. As a result, we are able to get
hold of decorated functions before they "disappear" inside an arbitrary
data structure of a custom decorator.

## Performance Considerations

The extra accuracy comes with memory and computing costs. The captured
code objects have to be stored on the module. We expect the memory cost
to be < 1% the overall memory footprint of the target application. When
a function location is looked up, the code object needs to be resolved
to a function. We query the GC to find what we are looking for. This is
generally an expensive operation, so we perform it on demand, when an
instrumentation needs to be applied, instead of looking up every code
object after a module import.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)