fix(datastreams): botocore - log warning on kinesis stream metadata not found (#11647)
For Data Streams Monitoring with AWS Kinesis, the map on the DSM page
will break if `StreamARN` isn't provided to the Kinesis `get_records`
call. This is because dd-trace-py requires the `StreamARN` when
generating a consume checkpoint for an in-edge in the data streams map.
If there's no consume checkpoint, the data streams map incorrectly
renders as two graphs instead of one connected graph.

However, the Kinesis `get_records` API lists `StreamARN` as an optional
parameter (see
[docs](https://boto3.amazonaws.com/v1/documentation/api/1.35.9/reference/services/kinesis/client/get_records.html)).
If it isn't provided, dd-trace-py only outputs a debug-level log, and
the DSM map is rendered incorrectly. This PR ensures dd-trace-py outputs
a warning that is more visible to the developer.

This PR can be closed if the current state of error handling is
acceptable, since I don't know if there are restrictions on logging
warnings for this repo. The warning helps to debug the broken DSM map,
which is time-consuming to debug otherwise.

No testing updates are required since there isn't a unit test file for
this specific module.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)