chore(llmobs): refactor trace processor tests
The trace processor tests intermingled business logic with the implementation
of the trace processor. To separate them out, we introduce a few testing
fixtures useful for dealing with tracing and capturing llm obs events. This
reduces the amount of mocking to zero and allows us to test more realistically.

There's a bit of a nasty hack to make sure the configs are updated in llmobs
modules that grab a reference to it but I'll follow up to clean that up as well.