Intentionally leak mutex
Refactor the postfork handler in the child to use placement-new in order
to leak the old mutex and create a new one.  We can't destroy, unlock,
reinitialize, or reset because all of those operations are UB in the
standard.  We could reimplement the locking mechanisms on top of
pthreads, but then in our port to Windows/macos, we'd need to replace
those primitives with equivalents on the respective platforms.

At the same time, the typical Python application which calls `fork()`
only does so one level deep--in other words, it is common for a
coordinator process to fork one level of children, but it is
unbelievably uncommon for those children to have long-lived forks of
their own (`popen()` sure, but absolutely no workers spawning workers).