chore(llmobs): refactor trace processor tests (#11784)
The trace processor tests intermingled business logic with the
implementation of the trace processor. To separate them out, we
introduce a few testing fixtures useful for dealing with tracing and
capturing llm obs events. This reduces the amount of mocking to zero and
allows us to test more realistically.

There's a bit of a nasty hack to make sure the configs are updated in
llmobs modules that grab a reference to it but I'll follow up to clean
that up as well.

This reduces most of the tests from this:

```python
def test_input_parameters_are_set():
    """Test that input parameters are set on the span event if they are present on the span."""
    dummy_tracer = DummyTracer()
    mock_llmobs_span_writer = mock.MagicMock()
    with override_global_config(dict(_llmobs_ml_app="unnamed-ml-app")):
        with dummy_tracer.trace("root_llm_span", span_type=SpanTypes.LLM) as llm_span:
            llm_span._set_ctx_item(SPAN_KIND, "llm")
            llm_span._set_ctx_item(INPUT_PARAMETERS, {"key": "value"})
        tp = LLMObsTraceProcessor(llmobs_span_writer=mock_llmobs_span_writer)
        assert tp._llmobs_span_event(llm_span)[0]["meta"]["input"]["parameters"] == {"key": "value"}
```
to this:

```python
def test_input_parameters_are_set(tracer, llmobs_events):
    """Test that input parameters are set on the span event if they are present on the span."""
    with tracer.trace("root_llm_span", span_type=SpanTypes.LLM) as llm_span:
        llm_span._set_ctx_item(const.SPAN_KIND, "llm")
        llm_span._set_ctx_item(const.INPUT_PARAMETERS, {"key": "value"})
    assert llmobs_events[0]["meta"]["input"]["parameters"] == {"key": "value"}
```