refactor(profiling): add debug check that the GIL is held to memalloc (#11860)
The allocation profiler functions should (in theory) only be called when
allocating Python objects, which should (in theory) only be done with
the GIL held. We have reason to believe that this code is sometimes
reached without the GIL held, since some crashes in the code seem to go
away with our own locking. Add a debug flag,
`_DD_PROFILING_MEMALLOC_CRASH_ON_NO_GIL`, that when set will make the
profiler crash if it detects the GIL is not held in places where we
think it ought to be.