fix(llmobs): replace trace processor with event listener (#11781)
The LLMObs service formerly depended on the TraceProcessor interface in
the tracer. This was problematic due to sharing a dependency with the
public API. As such, users could configure a trace filter (under the
hood is a trace processor) and overwrite the LLMObs TraceProcessor.

Instead, the tracer can emit span start and finish events which the
LLMObs service listens to and acts on, as proposed here.

The gotcha is that the LLMObs service no longer has a way to drop traces
when run in agentless mode, which only LLMObs supports. Instead, we
encourage users to explicitly turn off APM which carries the benefit of
clarity since this was implicit before.

Co-authored-by: Yun Kim <yun.kim@datadoghq.com>