feat(propagation): add DD_TRACE_PROPAGATION_BEHAVIOR_EXTRACT to handle x-org propagation (#11631)
Configuration: DD_TRACE_PROPAGATION_BEHAVIOR_EXTRACT
Values:
- continue (default): The tracing library continues the trace from the
incoming headers, if present. Also, incoming baggage is propagated.
- restart: The tracing library always starts a new trace with a new
trace-id. If there are distributed tracing headers, a span link will be
added to reference the trace context. Also, incoming baggage is
propagated.
- ignore: The tracing library always starts a new trace with a new
trace-id without creating any span links. Also, incoming baggage is
dropped.
Note: We do not need to implement this at the moment because the
workaround we have in place DD_TRACE_PROPAGATION_STYLE_EXTRACT=none
performs this exact function of completely ignoring incoming trace
context headers and baggage headers. If we receive feedback, we can add
this new option that we maintain.

Result
✅ Head services: Customers can configure these services with the restart
configuration so that they are always the root span, regardless of
incoming distributed tracing headers.
✅ Internal services: By default, internal services will remain unchanged
and continue to report complete traces to Datadog. If a user has
configured an internal service with the restart configuration, this will
lead to partial traces. Thankfully with Span Links we can easily direct
the customer to the upstream trace (if it was sampled).
❌ Sometimes-head services: This solution does not solve for services
that are sometimes head services and sometimes not. The best solution
for these services is to set the restart configuration so that a new
trace can begin and sampling priority can be recalculated, and in the
case that this was not a head service then the upstream service can be
found with a span link.
❌ Reentrant systems: This solution does not improve this scenario no
matter how the reentrant service is configured. Under the continue
configuration, the reentrant service is part of the same trace as the
original root span but it is orphaned. Under the restart configuration,
a new trace will be started and the span will be decorated with a span
link containing the original trace-id but the contained span-id would
not point to a span tracked in their Datadog organization.


Note: This contains a fix, allowing only baggage to be propagated.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: erikayasuda <153395705+erikayasuda@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>
Co-authored-by: Christophe Papazian <114495376+christophe-papazian@users.noreply.github.com>
Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>