chore(llmobs): implement ragas context precision (#11716)
Implements context precision metric for ragas integration.

## About context precision
[Context
Precision](https://docs.ragas.io/en/latest/concepts/metrics/available_metrics/context_precision/)
is a metric that verifies if the context was useful in arriving at the
given answer.
We compute this by dividing the number of relevant contexts by the total
number of contexts.
Note that this is slightly modified from the original context precision
metric in ragas, which computes
the mean of the precision @ rank k for each chunk in the context (where
k is the number of
        retrieved context chunks).

**stage 1 - input extraction**
We must extract a question, contexts, and answer from an LLM span. 
_Question_ - same as faithfulness (last input message or `question`
prompt variable)
_Contexts_ - uses a **list** of context texts that ground the
generation. Defaults to just the context prompt variable being the
single item in the list or user defined prompt context variables
_Answer_ - first output message for the llm span

**stage 2 - prompt creation / llm calls**
For each context, create a prompt that asks an LLM if the context was
used in the generated answer & execute llm calls for these prompts.

**stage 3 - score computation**
Divide the number of "used" contexts by total number of contexts

**why do we modify?**
The ragas context precision metric is a mean of context precision @ k,
where k is number of chunks returned. This uses number of relevant
context docs as the denominator.

We only have access to the context information on an LLM call and not at
the retrieval step. For most users, there should only be one "context"
being used to ground the LLM call. So if we use the original ragas
formula, the denominator may frequently be zero if the LLM decided the
context was not used. Thus, we revise the metric to answer a simple
question:
- was my context used in the llm's generation (in the case of a single
context)
- what proportion of my contexts were used in the llm's generation (in
the case of multiple contexts)

### Example trace
<img width="1127" alt="image"
src="https://github.com/user-attachments/assets/d6ea803b-7511-4720-ad1e-6230cc35fdb5"
/>


## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: lievan <evan.li@datadoqhq.com>
Co-authored-by: Quinna Halim <quinna.halim@datadoghq.com>