fix(tracing): only extract distributed headers if a trace is not already started (#9456)
Today there are some integrations which always check for distributed
tracing headers and if found will activate them as the current/parent
context. This blind approach does not take into account if a parent
trace has already been started.

This means if you have nested Flask apps (for example), then the child
Flask app may make it's parent the upstream remote service, and not the
first/initial Flask app in the process. This can result in a broken
trace.

The scenario which "works" today which will be broken with this change
is if there is "broken" instrumentation as the first span in a trace.
This could either be custom instrumentation that did not check for
distributed headers, or a built-in integration which is not properly
parsing/activating distributed tracing headers.

This change would prioritize keeping the local trace accurate over
ensuring a connection with the upstream calling service.


## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [ ] Title is accurate
- [ ] All changes are related to the pull request's stated goal
- [ ] Description motivates each change
- [ ] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [ ] Testing strategy adequately addresses listed risks
- [ ] Change is maintainable (easy to change, telemetry, documentation)
- [ ] Release note makes sense to a user of the library
- [ ] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [ ] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Federico Mon <federico.mon@datadoghq.com>
Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>
Co-authored-by: Andrew Lock <andrewlock.net@gmail.com>
Co-authored-by: Zachary Groves <32471391+ZStriker19@users.noreply.github.com>
Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>
Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>
Co-authored-by: Christophe Papazian <114495376+christophe-papazian@users.noreply.github.com>