feat(langgraph): submit spans from langgraph to APM and LLMObs (#11730)
This PR adds a simple integration and corresponding LLMObs plugin for
the LangGraph framework.

### Enablement

For both APM and LLMObs, this integration is disabled by default, and
only enabled with `DD_TRACE_LANGGRAPH_ENABLED=true` or
`patch(langgraph=True)`, along with `_DD_TRACE_LANGGRAPH_ENABLED=true`,
due to its experimental nature.

### APM Tracing

The APM integration adds basic tracing for LangGraph `Pregel` graph
invocations, as well as the execution of each node/step in the graph
(traced as `RunnableSeq`).

### LLM Observability Plugin

The LLMObs plugin submits the traces to LLM Observability.

#### Graph Invocation/Streaming Output Tagging

Because we patch the `(a)stream` method for both streams and
`(a)invoke`, since they rely on `(a)stream` under the hood, we are bound
to the yields from the stream. The object shape and value depend on the
`stream_mode` passed in to the stream function. For `invoke`, it
defaults to `"values"`, which is perfect for our use case. However, if
one were to call `stream` directly, and either pass in nothing or
something like `stream_mode=["updates", "debug"]`, the output we would
have for the graph span would appear as:

```
{
  "type": "task_result",
  "timestamp": "2025-01-03T14:45:26.087608+00:00",
  "step": "3",
  "payload": {
    "id": "399cfe0a-aefd-f67c-6909-fcea80ef508b",
    "name": "e",
    "error": "None",
    "result": [
      "('a_list', ['e'])"
    ],
    "interrupts": []
  }
}
```

This is a known limitation for now, and a more thought-out solution for
either adding on `"values"` to `stream_mode` and omitting their yields
if it wasn't already present, or having some kind of clever joining, can
be added in a later PR or by feature request.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Yun Kim <yun.kim@datadoghq.com>
Co-authored-by: Yun Kim <35776586+Yun-Kim@users.noreply.github.com>